pyproject.toml
[project]
name = "salesforce-data-modeler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=1.2.0",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "simple-salesforce>=1.12.6",
    "xlsxwriter>=3.2.2",
]
